requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements-dev.txt
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via starlette
black==24.8.0
    # via resource-catalogue-fastapi (pyproject.toml)
boto3==1.35.38
    # via resource-catalogue-fastapi (pyproject.toml)
botocore==1.35.38
    # via
    #   boto3
    #   s3transfer
build==1.2.2
    # via pip-tools
certifi==2024.8.30
    # via
    #   pulsar-client
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
    #   uvicorn
distlib==0.3.8
    # via virtualenv
execnet==2.1.1
    # via pytest-xdist
fastapi==0.115.0
    # via resource-catalogue-fastapi (pyproject.toml)
filelock==3.16.1
    # via virtualenv
h11==0.14.0
    # via uvicorn
identify==2.6.1
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   requests
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via resource-catalogue-fastapi (pyproject.toml)
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mypy-extensions==1.0.0
    # via black
nodeenv==1.9.1
    # via pre-commit
packaging==24.1
    # via
    #   black
    #   build
    #   pytest
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via resource-catalogue-fastapi (pyproject.toml)
platformdirs==4.3.6
    # via
    #   black
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==3.8.0
    # via resource-catalogue-fastapi (pyproject.toml)
pulsar-client==3.5.0
    # via resource-catalogue-fastapi (pyproject.toml)
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
pyjwt==2.9.0
    # via resource-catalogue-fastapi (pyproject.toml)
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.3
    # via
    #   pytest-mock
    #   pytest-xdist
    #   resource-catalogue-fastapi (pyproject.toml)
pytest-mock==3.14.0
    # via resource-catalogue-fastapi (pyproject.toml)
pytest-watcher==0.4.3
    # via resource-catalogue-fastapi (pyproject.toml)
pytest-xdist==3.6.1
    # via resource-catalogue-fastapi (pyproject.toml)
python-dateutil==2.9.0.post0
    # via botocore
pyyaml==6.0.2
    # via
    #   pre-commit
    #   resource-catalogue-fastapi (pyproject.toml)
requests==2.32.3
    # via resource-catalogue-fastapi (pyproject.toml)
ruff==0.6.8
    # via resource-catalogue-fastapi (pyproject.toml)
s3transfer==0.10.3
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.38.6
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.2.3
    # via
    #   botocore
    #   requests
uvicorn==0.31.1
    # via resource-catalogue-fastapi (pyproject.toml)
virtualenv==20.26.6
    # via pre-commit
watchdog==5.0.3
    # via pytest-watcher
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
